# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/applications/url": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Create From Job Url",
        "operationId": "applications-create_from_job_url",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationByJobUrl"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/web__applications__api_schema__Application"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Oauth2PasswordBearerFromCookies": []
          }
        ]
      }
    },
    "/applications/status_flow": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Get Allowed Statuses For Update",
        "operationId": "applications-get_allowed_statuses_for_update",
        "security": [
          {
            "Oauth2PasswordBearerFromCookies": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ApplicationStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationStatus"
                  },
                  "title": "Response Applications-Get Allowed Statuses For Update"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/applications/{application_id}": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Get Application",
        "operationId": "applications-get_application",
        "security": [
          {
            "Oauth2PasswordBearerFromCookies": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/web__applications__api_schema__Application"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "applications"
        ],
        "summary": "Application Partial Update",
        "operationId": "applications-application_partial_update",
        "security": [
          {
            "Oauth2PasswordBearerFromCookies": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationPartial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/lib__applications__models__Application"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/applications/": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "List Applications",
        "operationId": "applications-list_applications",
        "security": [
          {
            "Oauth2PasswordBearerFromCookies": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ApplicationSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "Pending",
          "Applied",
          "Received by Employer",
          "In Assessment",
          "In Screening",
          "Interviewing",
          "Offer Received",
          "Negotiating",
          "Offer Declined",
          "Hired",
          "Rejected",
          "Ghosted"
        ],
        "title": "ApplicationStatus"
      },
      "ApplicationSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "fit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fit"
          },
          "job": {
            "$ref": "#/components/schemas/JobSummary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "job"
        ],
        "title": "ApplicationSummary"
      },
...
      "Company": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug",
          "id"
        ],
        "title": "Company"
      },
      "Compensation": {
        "properties": {
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "hiring_bonus": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hiring Bonus"
          },
          "equity": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equity"
          },
          "minimum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum"
          },
          "maximum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximum"
          },
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "benefits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Benefits",
            "default": []
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "Compensation"
      },
      "CompensationSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "minimum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum"
          },
          "maximum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximum"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "CompensationSummary"
      },
      "CreateApplicationByJobUrl": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "CreateApplicationByJobUrl"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Job": {
        "properties": {
          "job_title": {
            "type": "string",
            "title": "Job Title"
          },
          "job_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Job Url"
          },
          "job_type": {
            "type": "string",
            "enum": [
              "full_time",
              "part_time",
              "contract"
            ],
            "title": "Job Type"
          },
          "llm_summary": {
            "type": "string",
            "title": "Llm Summary"
          },
          "job_description": {
            "type": "string",
            "title": "Job Description"
          },
          "requirements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requirements",
            "default": []
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          },
          "modality": {
            "items": {
              "type": "string",
              "enum": [
                "remote",
                "in_office",
                "hybrid"
              ]
            },
            "type": "array",
            "title": "Modality",
            "default": []
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "other_details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Details"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "compensation": {
            "$ref": "#/components/schemas/Compensation"
          },
          "company": {
            "$ref": "#/components/schemas/Company"
          }
        },
        "type": "object",
        "required": [
          "job_title",
          "job_url",
          "job_type",
          "llm_summary",
          "job_description",
          "location",
          "id",
          "compensation",
          "company"
        ],
        "title": "Job"
      },
      "JobSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "job_title": {
            "type": "string",
            "title": "Job Title"
          },
          "job_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Job Url"
          },
          "llm_summary": {
            "type": "string",
            "title": "Llm Summary"
          },
          "modality": {
            "items": {
              "type": "string",
              "enum": [
                "remote",
                "in_office",
                "hybrid"
              ]
            },
            "type": "array",
            "title": "Modality"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "compensation": {
            "$ref": "#/components/schemas/CompensationSummary"
          },
          "company": {
            "$ref": "#/components/schemas/Company"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_title",
          "job_url",
          "llm_summary",
          "modality",
          "location",
          "compensation",
          "company"
        ],
        "title": "JobSummary"
      },
      "PaginatedResponse_ApplicationSummary_": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "next_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Page"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "item_count": {
            "type": "integer",
            "title": "Item Count"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ApplicationSummary"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "page",
          "next_page",
          "total_pages",
          "item_count",
          "data"
        ],
        "title": "PaginatedResponse[ApplicationSummary]"
      },
...
      "UpdateApplicationPartial": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "interview_rounds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Rounds"
          },
          "current_round": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Round"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "fit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fit"
          },
          "resume_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Used"
          }
        },
        "type": "object",
        "title": "UpdateApplicationPartial"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "lib__applications__models__Application": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus",
            "default": "Pending"
          },
          "interview_rounds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Rounds"
          },
          "current_round": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Round"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "fit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fit"
          },
          "resume_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Used"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          }
        },
        "type": "object",
        "title": "Application"
      },
      "web__applications__api_schema__Application": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus",
            "default": "Pending"
          },
          "interview_rounds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Rounds"
          },
          "current_round": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Round"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "fit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fit"
          },
          "resume_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Used"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "job": {
            "$ref": "#/components/schemas/Job"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job"
        ],
        "title": "Application"
      }
    },
    "securitySchemes": {
      "Oauth2PasswordBearerFromCookies": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}